# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /uploads/{category}/{filename}:
    get:
      operationId: serveProjectUpload
      tags: [admin]
      summary: Serve an uploaded avatar or window-profile image
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
            enum: [avatars, window-profiles]
        - name: filename
          in: path
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9-]+\\.(jpg|png|webp)$"
      responses:
        "200":
          description: Uploaded image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "401":
          description: Sign-in is required
        "403":
          description: Required module access is missing
        "404":
          description: Image not found
  /users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List users for the administration screen
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: roleId
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a user with an administrator-assigned password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Edit a user, permissions, or password
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deactivateUser
      tags: [admin]
      summary: Deactivate a user without deleting audit history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deactivated
  /roles:
    get:
      operationId: listRoles
      tags: [admin]
      summary: List configurable roles and permission templates
      responses:
        "200":
          description: Role records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [admin]
      summary: Create a custom role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{roleId}:
    patch:
      operationId: updateRole
      tags: [admin]
      summary: Edit a role and its default permissions
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      operationId: deleteRole
      tags: [admin]
      summary: Delete an unused custom role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Role deleted
...
components:
  schemas:
...
    PermissionLevel:
      type: string
      enum: [none, view, edit]
    PermissionMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/PermissionLevel"
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isSystem:
          type: boolean
        permissions:
          $ref: "#/components/schemas/PermissionMap"
      required: [id, name, description, isSystem, permissions]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        roleId:
          type: string
        roleName:
          type: string
        status:
          type: string
          enum: [active, inactive]
        lastLogin:
          type: ["string", "null"]
          format: date-time
        permissions:
          $ref: "#/components/schemas/PermissionMap"
        permissionOverrides:
          type: ["object", "null"]
          additionalProperties:
            $ref: "#/components/schemas/PermissionLevel"
      required:
        - id
        - name
        - username
        - email
        - phone
        - avatarUrl
        - roleId
        - roleName
        - status
        - lastLogin
        - permissions
        - permissionOverrides
...
    LoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required: [username, password]
    UserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 3
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
          maxLength: 180000
        roleId:
          type: string
        password:
          type: string
          minLength: 8
        permissionOverrides:
          type: ["object", "null"]
          additionalProperties:
            $ref: "#/components/schemas/PermissionLevel"
      required: [name, username, roleId, password]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 3
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
          maxLength: 180000
        roleId:
          type: string
        password:
          type: string
          minLength: 8
        status:
          type: string
          enum: [active, inactive]
        permissionOverrides:
          type: ["object", "null"]
          additionalProperties:
            $ref: "#/components/schemas/PermissionLevel"
    RoleInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: ["string", "null"]
        permissions:
          $ref: "#/components/schemas/PermissionMap"
      required: [name, permissions]
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: ["string", "null"]
        permissions:
          $ref: "#/components/schemas/PermissionMap"
...